Move kubeadm upstream types to upstream<suffix>
Having the same package names was causing conflicts in the conversion
generation code, for example InitConfiguration is in both CABPK's code
and upstream types.
