+++ bootstrap/kubeadm/types/upstreamv1beta1/bootstraptokenstring_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 The Kubernetes Authors.
+Copyright 2018 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1beta2
+package upstreamv1beta1
 
 import (
 	"encoding/json"
@@ -41,7 +41,7 @@ func TestMarshalJSON(t *testing.T) {
 
 			b, err := json.Marshal(rt.bts)
 			g.Expect(err).NotTo(HaveOccurred())
-			g.Expect(string(b)).To(Equal(rt.expected))
+			g.Expect(b).To(BeEquivalentTo(rt.expected))
 		})
 	}
 }
