+++ bootstrap/kubeadm/types/upstreamv1beta2/doc.go
@@ -14,11 +14,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package v1beta1 contains a mirror of kubeadm API v1beta1 API, required because it is not possible to import k/K.
+// Package upstreamv1beta2 contains a mirror of kubeadm API v1beta2 API, required because it is not possible to import k/K.
 //
 // IMPORTANT: Do not change these files!
 // IMPORTANT: only for KubeadmConfig serialization/deserialization, and should not be used for other purposes.
 //
-// +k8s:conversion-gen=sigs.k8s.io/cluster-api/bootstrap/kubeadm/api/v1alpha4
+// +k8s:conversion-gen=sigs.k8s.io/cluster-api/bootstrap/kubeadm/api/v1beta1
 // +k8s:deepcopy-gen=package
-package v1beta1 // import "sigs.k8s.io/cluster-api/bootstrap/kubeadm/kubeadm/v1beta1"
+package upstreamv1beta2 // import "sigs.k8s.io/cluster-api/bootstrap/kubeadm/types/upstreamv1beta2"
