+++ bootstrap/kubeadm/types/upstreamv1beta2/groupversion_info.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v1beta2
+package upstreamv1beta2
 
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
