+++ bootstrap/kubeadm/types/utils.go
@@ -23,10 +23,10 @@ import (
 	runtime "k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/runtime/serializer"
-	bootstrapv1 "sigs.k8s.io/cluster-api/bootstrap/kubeadm/api/v1alpha4"
-	"sigs.k8s.io/cluster-api/bootstrap/kubeadm/types/v1beta1"
-	"sigs.k8s.io/cluster-api/bootstrap/kubeadm/types/v1beta2"
-	"sigs.k8s.io/cluster-api/bootstrap/kubeadm/types/v1beta3"
+	bootstrapv1 "sigs.k8s.io/cluster-api/bootstrap/kubeadm/api/v1beta1"
+	"sigs.k8s.io/cluster-api/bootstrap/kubeadm/types/upstreamv1beta1"
+	"sigs.k8s.io/cluster-api/bootstrap/kubeadm/types/upstreamv1beta2"
+	"sigs.k8s.io/cluster-api/bootstrap/kubeadm/types/upstreamv1beta3"
 	"sigs.k8s.io/controller-runtime/pkg/conversion"
 	"sigs.k8s.io/controller-runtime/pkg/scheme"
 )
@@ -37,27 +37,27 @@ var (
 	v1beta3KubeadmVersion = semver.MustParse("1.22.0")
 
 	clusterConfigurationVersionTypeMap = map[schema.GroupVersion]conversion.Convertible{
-		v1beta3.GroupVersion: &v1beta3.ClusterConfiguration{},
-		v1beta2.GroupVersion: &v1beta2.ClusterConfiguration{},
-		v1beta1.GroupVersion: &v1beta1.ClusterConfiguration{},
+		upstreamv1beta3.GroupVersion: &upstreamv1beta3.ClusterConfiguration{},
+		upstreamv1beta2.GroupVersion: &upstreamv1beta2.ClusterConfiguration{},
+		upstreamv1beta1.GroupVersion: &upstreamv1beta1.ClusterConfiguration{},
 	}
 
 	clusterStatusVersionTypeMap = map[schema.GroupVersion]conversion.Convertible{
 		// ClusterStatus has been removed in v1beta3, so we don't need an entry for v1beta3
-		v1beta2.GroupVersion: &v1beta2.ClusterStatus{},
-		v1beta1.GroupVersion: &v1beta1.ClusterStatus{},
+		upstreamv1beta2.GroupVersion: &upstreamv1beta2.ClusterStatus{},
+		upstreamv1beta1.GroupVersion: &upstreamv1beta1.ClusterStatus{},
 	}
 
 	initConfigurationVersionTypeMap = map[schema.GroupVersion]conversion.Convertible{
-		v1beta3.GroupVersion: &v1beta3.InitConfiguration{},
-		v1beta2.GroupVersion: &v1beta2.InitConfiguration{},
-		v1beta1.GroupVersion: &v1beta1.InitConfiguration{},
+		upstreamv1beta3.GroupVersion: &upstreamv1beta3.InitConfiguration{},
+		upstreamv1beta2.GroupVersion: &upstreamv1beta2.InitConfiguration{},
+		upstreamv1beta1.GroupVersion: &upstreamv1beta1.InitConfiguration{},
 	}
 
 	joinConfigurationVersionTypeMap = map[schema.GroupVersion]conversion.Convertible{
-		v1beta3.GroupVersion: &v1beta3.JoinConfiguration{},
-		v1beta2.GroupVersion: &v1beta2.JoinConfiguration{},
-		v1beta1.GroupVersion: &v1beta1.JoinConfiguration{},
+		upstreamv1beta3.GroupVersion: &upstreamv1beta3.JoinConfiguration{},
+		upstreamv1beta2.GroupVersion: &upstreamv1beta2.JoinConfiguration{},
+		upstreamv1beta1.GroupVersion: &upstreamv1beta1.JoinConfiguration{},
 	}
 )
 
@@ -68,10 +68,10 @@ func KubeVersionToKubeadmAPIGroupVersion(version semver.Version) (schema.GroupVe
 		return schema.GroupVersion{}, errors.New("the bootstrap provider for kubeadm doesn't support Kubernetes version lower than v1.13.0")
 	case version.LT(v1beta2KubeadmVersion):
 		// NOTE: All the Kubernetes version >= v1.13 and < v1.15 should use the kubeadm API version v1beta1
-		return v1beta1.GroupVersion, nil
+		return upstreamv1beta1.GroupVersion, nil
 	case version.LT(v1beta3KubeadmVersion):
 		// NOTE: All the Kubernetes version >= v1.15 and < v1.22 should use the kubeadm API version v1beta2
-		return v1beta2.GroupVersion, nil
+		return upstreamv1beta2.GroupVersion, nil
 	default:
 		// NOTE: All the Kubernetes version greater or equal to v1.22 should use the kubeadm API version v1beta3.
 		// Also future Kubernetes versions (not yet released at the time of writing this code) are going to use v1beta3,
@@ -79,7 +79,7 @@ func KubeVersionToKubeadmAPIGroupVersion(version semver.Version) (schema.GroupVe
 		// This is acceptable because v1beta3 will be supported by kubeadm until the deprecation cycle completes
 		// (9 months minimum after the deprecation date, not yet announced now); this gives Cluster API project time to
 		// introduce support for newer releases without blocking users to deploy newer version of Kubernetes.
-		return v1beta3.GroupVersion, nil
+		return upstreamv1beta3.GroupVersion, nil
 	}
 }
